Not much changes fundamentally
In the csv file, I included the headers this time so that it is easier to read when not accessed in Excel. I changed the table name in the sql file to some more shorter and comprehensive. There's an issue where psql says it's missing data for column "location", but that's due to the fact that there are some columns left empty. If anyone knows how to circumvent that or has any ideas, let me know.