File Changes and Updates
I have once again done a fresh download of the March 11, 2019 data so that it has all of the requests for that day and not just the ones made at midnight. I have also update the dataload file so that it creates the filter tables necessary for certain requests like order by geographic location and which cases are resolved or unresolved.