Updated with more filter tables
I added a search/order by case type filter, number of calls, and search by time of day sample query. I made it a sample query to show how to do it because a time of day search is relative to what time the person wants. All the data is available via the database already.